Make sure mobile-related attributes are reset at every mobile selection (TID)
Since we are skipping the call to `Claim#reset_dependent_answers` (ref .1) when firing the update
action from `select-mobile`  – and this is necessary because we update the mobile-related attributes
using a custom pseudo-form object (ref. 2) – we need a way to reset the `mobile_verified` attribute
that could be set to `true` if the user decides to provide and confirm an alternative number.

This is an issue when changing answers, as highlighted by the additional scenarios added to the
relevant feature tests.

This commit makes sure that `mobile_verified` is always `nil` on the TID route when firing an
update from `select-mobile` so to make sure the correct slug is presented when changing answers.

ref 1: https://github.com/DFE-Digital/claim-additional-payments-for-teaching/blob/a723265ef0e9f60b489a939f2438997a37b6bc2c/app/controllers/claims_controller.rb#L110
ref 2: https://github.com/DFE-Digital/claim-additional-payments-for-teaching/blob/a723265ef0e9f60b489a939f2438997a37b6bc2c/app/forms/select_mobile_number_form.rb